+++ functions/Config.fnc.php
@@ -33,3 +33,40 @@ function Config( $item )
 
 	return $_ROSARIO['Config'][$item][1]['CONFIG_VALUE'];
 }
+
+
+/**
+ * Get Program Configuration
+ * Get 1 value if item specified,
+ * else get Program values
+ *
+ * Values set in School Configuration or directly in Module (ex.: Eligibility Entry times)
+ *
+ * @example if ( ProgramConfig( 'STUDENTS_SEMESTER_COMMENTS', 'students' ) )
+ *
+ * @since 2.9
+ *
+ * @param  string       $program eligibility|grades|students|moodle|food_service|attendance... Program name
+ * @param  string       $item    Program Config title (optional). Defaults to 'all'
+ *
+ * @return string|array Program Configuration value, or Program values in array
+ */
+function ProgramConfig( $program, $item = 'all'  )
+{
+	global $_ROSARIO;
+
+	if ( !isset( $_ROSARIO['ProgramConfig'] ) )
+	{
+		$_ROSARIO['ProgramConfig'] = DBGet( DBQuery( "SELECT PROGRAM,TITLE,VALUE
+			FROM PROGRAM_CONFIG
+			WHERE SYEAR='" . UserSyear() . "'
+			AND SCHOOL_ID='" . UserSchool() . "'" ), array(), array( 'PROGRAM', 'TITLE' ) );
+	}
+
+	if ( $item === 'all' )
+	{
+		return $_ROSARIO['ProgramConfig'][$program];
+	}
+	else
+		return $_ROSARIO['ProgramConfig'][$program][$item][1]['VALUE'];
+}