+++ modules/Eligibility/EnterEligibility.php
@@ -1,9 +1,10 @@
 <?php
-$start_end_RET = DBGet(DBQuery("SELECT TITLE,VALUE FROM PROGRAM_CONFIG WHERE SYEAR='".UserSyear()."' AND SCHOOL_ID='".UserSchool()."' AND PROGRAM='eligibility'"));
-if(count($start_end_RET))
+// GET ALL THE CONFIG ITEMS FOR ELIGIBILITY
+$eligibility_config = ProgramConfig( 'eligibility' );
+
+foreach( (array)$eligibility_config as $value )
 {
-	foreach($start_end_RET as $value)
-		$$value['TITLE'] = $value['VALUE'];
+	${$value[1]['TITLE']} = $value[1]['VALUE'];
 }
 
 switch(date('D'))