+++ modules/Eligibility/Student.php
@@ -31,11 +31,12 @@ if($_REQUEST['modfunc']=='remove' && AllowEdit())
 
 if(UserStudentID() && !$_REQUEST['modfunc'])
 {
-	$start_end_RET = DBGet(DBQuery("SELECT TITLE,VALUE FROM PROGRAM_CONFIG WHERE SYEAR='".UserSyear()."' AND SCHOOL_ID='".UserSchool()."' AND PROGRAM='eligibility' AND TITLE IN ('START_DAY','END_DAY')"));
-	if(count($start_end_RET))
+	// GET ALL THE CONFIG ITEMS FOR ELIGIBILITY
+	$eligibility_config = ProgramConfig( 'eligibility' );
+
+	foreach( (array)$eligibility_config as $value )
 	{
-		foreach($start_end_RET as $value)
-			$$value['TITLE'] = $value['VALUE'];
+		${$value[1]['TITLE']} = $value[1]['VALUE'];
 	}
 	
 	switch(date('D'))
@@ -139,5 +140,3 @@ function _makeLower($word)
 {
 	return ucwords(mb_strtolower($word));
 }
-
-?>