+++ modules/Grades/Configuration.php
@@ -1,6 +1,4 @@
 <?php
-//FJ add School Configuration
-$program_config = DBGet(DBQuery("SELECT * FROM PROGRAM_CONFIG WHERE SCHOOL_ID='".UserSchool()."' AND SYEAR='".UserSyear()."' AND PROGRAM='grades'"),array(),array('TITLE'));
 
 if($_REQUEST['values'])
 {