+++ modules/School_Setup/Configuration.php
@@ -47,8 +47,9 @@ else
 				if (isset($_REQUEST['values']['CONFIG']) && is_array($_REQUEST['values']['CONFIG']))
 					foreach($_REQUEST['values']['CONFIG'] as $column=>$value)
 					{
-						$sql .= "UPDATE CONFIG SET ";
-						$sql .= "CONFIG_VALUE='".$value."' WHERE TITLE='".$column."'";
+						$sql .= "UPDATE CONFIG SET
+							CONFIG_VALUE='" . $value . "'
+							WHERE TITLE='" . $column . "'";
 					
 						//Program Title, Program Name, Default Theme, Create User Account, Create Student Account
 						$school_independant_values = array('TITLE','NAME','THEME', 'CREATE_USER_ACCOUNT', 'CREATE_STUDENT_ACCOUNT');
@@ -56,24 +57,29 @@ else
 						if (in_array($column,$school_independant_values))
 							$sql .= " AND SCHOOL_ID='0';";
 						else
-							$sql .= " AND SCHOOL_ID='".UserSchool()."';";
+							$sql .= " AND SCHOOL_ID='" . UserSchool() . "';";
 					}
 
 				if (isset($_REQUEST['values']['PROGRAM_CONFIG']) && is_array($_REQUEST['values']['PROGRAM_CONFIG']))
 					foreach($_REQUEST['values']['PROGRAM_CONFIG'] as $column=>$value)
 					{
-						$sql .= "UPDATE PROGRAM_CONFIG SET ";
-						$sql .= "VALUE='".$value."' WHERE TITLE='".$column."'";
-						$sql .= " AND SCHOOL_ID='".UserSchool()."' AND SYEAR='".UserSyear()."';";
+						$sql .= "UPDATE PROGRAM_CONFIG SET 
+							VALUE='" . $value . "'
+							WHERE TITLE='" . $column . "'
+							AND SCHOOL_ID='" . UserSchool() . "'
+							AND SYEAR='" . UserSyear() . "';";
 					}
 
 				if ($sql != '')
 				{
 					DBQuery($sql);
+
 					$note[] = button('check') .'&nbsp;'._('The school configuration has been modified.');
 				}
-				
-				unset($_ROSARIO['Config']);//update Config var
+
+				unset( $_ROSARIO['Config'] ); // update Config var
+
+				unset( $_ROSARIO['ProgramConfig'] ); // update ProgramConfig var
 			}
 			else
 			{
@@ -88,21 +94,19 @@ else
 
 	if(empty($_REQUEST['modfunc']))
 	{
+		echo '<FORM ACTION="Modules.php?modname='.$_REQUEST['modname'].'&modfunc=update" METHOD="POST" enctype="multipart/form-data">';
+
+		if(AllowEdit())
+			DrawHeader('',SubmitButton(_('Save')));
+
 		if (!empty($note))
 			echo ErrorMessage($note, 'note');
 
 		if (!empty($error))
 			echo ErrorMessage($error, 'error');
-		
-		echo '<FORM ACTION="Modules.php?modname='.$_REQUEST['modname'].'&modfunc=update" METHOD="POST" enctype="multipart/form-data" onsubmit="if (document.getElementById(\'LOGO_FILE\').value) document.getElementById(\'loading\').innerHTML=\'<span class=loading></span>\';">';
-	
-		if(AllowEdit())
-			DrawHeader('',SubmitButton(_('Save')));
-		
+
 		echo '<BR />';
 		PopTable('header',SchoolInfo('TITLE'));
-	
-		$program_config = DBGet(DBQuery("SELECT * FROM PROGRAM_CONFIG WHERE SCHOOL_ID='".UserSchool()."' AND SYEAR='".UserSyear()."'"),array(),array('TITLE'));
 	
 		echo '<FIELDSET><legend>'.ParseMLField(Config('TITLE')).'</legend><TABLE>';
 
@@ -156,11 +160,35 @@ else
 
 			echo '<TR><TD>'.CheckboxInput(Config('STUDENTS_USE_MAILING'), 'values[CONFIG][STUDENTS_USE_MAILING]',_('Display Mailing Address'), '', false, button('check'), button('x')).'</TD></TR>';
 
-			echo '<TR><TD>'.CheckboxInput($program_config['STUDENTS_USE_BUS'][1]['VALUE'], 'values[PROGRAM_CONFIG][STUDENTS_USE_BUS]', _('Check Bus Pickup / Dropoff by default'), '', false, button('check'), button('x')).'</TD></TR>';
+			echo '<TR><TD>'.CheckboxInput(
+				ProgramConfig( 'students', 'STUDENTS_USE_BUS' ),
+				'values[PROGRAM_CONFIG][STUDENTS_USE_BUS]',
+				_( 'Check Bus Pickup / Dropoff by default' ),
+				'',
+				false,
+				button( 'check' ),
+				button( 'x' )
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.CheckboxInput($program_config['STUDENTS_USE_CONTACT'][1]['VALUE'], 'values[PROGRAM_CONFIG][STUDENTS_USE_CONTACT]', _('Enable Legacy Contact Information'), '', false, button('check'), button('x')).'</TD></TR>';
+			echo '<TR><TD>'.CheckboxInput(
+				ProgramConfig( 'students', 'STUDENTS_USE_CONTACT' ),
+				'values[PROGRAM_CONFIG][STUDENTS_USE_CONTACT]',
+				_( 'Enable Legacy Contact Information' ),
+				'',
+				false,
+				button( 'check' ),
+				button( 'x' )
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.CheckboxInput($program_config['STUDENTS_SEMESTER_COMMENTS'][1]['VALUE'], 'values[PROGRAM_CONFIG][STUDENTS_SEMESTER_COMMENTS]', _('Use Semester Comments instead of Quarter Comments'), '', false, button('check'), button('x')).'</TD></TR>';
+			echo '<TR><TD>'.CheckboxInput(
+				ProgramConfig( 'students', 'STUDENTS_SEMESTER_COMMENTS' ),
+				'values[PROGRAM_CONFIG][STUDENTS_SEMESTER_COMMENTS]',
+				_( 'Use Semester Comments instead of Quarter Comments' ),
+				'',
+				false,
+				button( 'check' ),
+				button( 'x' )
+			) . '</TD></TR>';
 
 			echo '</TABLE></FIELDSET>';
 		}
@@ -170,15 +198,53 @@ else
 			echo '<BR /><FIELDSET><legend>'._('Grades').'</legend><TABLE>';
 			$options = array('-1' => _('Use letter grades only'), '0' => _('Use letter and percent grades'), '1' => _('Use percent grades only'));
 
-			echo '<TR><TD>'.SelectInput($program_config['GRADES_DOES_LETTER_PERCENT'][1]['VALUE'],'values[PROGRAM_CONFIG][GRADES_DOES_LETTER_PERCENT]',_('Grades'),$options,false).'</TD></TR>';
+			echo '<TR><TD>'.SelectInput(
+				ProgramConfig( 'grades', 'GRADES_DOES_LETTER_PERCENT' ),
+				'values[PROGRAM_CONFIG][GRADES_DOES_LETTER_PERCENT]',
+				_( 'Grades' ),
+				$options,
+				false
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.CheckboxInput($program_config['GRADES_HIDE_NON_ATTENDANCE_COMMENT'][1]['VALUE'], 'values[PROGRAM_CONFIG][GRADES_HIDE_NON_ATTENDANCE_COMMENT]', _('Hide grade comment except for attendance period courses'), '', false,button('check'), button('x')).'</TD></TR>';
+			echo '<TR><TD>'.CheckboxInput(
+				ProgramConfig( 'grades', 'GRADES_HIDE_NON_ATTENDANCE_COMMENT' ),
+				'values[PROGRAM_CONFIG][GRADES_HIDE_NON_ATTENDANCE_COMMENT]',
+				_( 'Hide grade comment except for attendance period courses' ),
+				'',
+				false,
+				button( 'check' ),
+				button( 'x' )
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.CheckboxInput($program_config['GRADES_TEACHER_ALLOW_EDIT'][1]['VALUE'], 'values[PROGRAM_CONFIG][GRADES_TEACHER_ALLOW_EDIT]', _('Allow Teachers to edit grades after grade posting period'), '', false, button('check'), button('x')).'</TD></TR>';
+			echo '<TR><TD>'.CheckboxInput(
+				ProgramConfig( 'grades', 'GRADES_TEACHER_ALLOW_EDIT' ),
+				'values[PROGRAM_CONFIG][GRADES_TEACHER_ALLOW_EDIT]',
+				_( 'Allow Teachers to edit grades after grade posting period' ),
+				'',
+				false,
+				button( 'check' ),
+				button( 'x' )
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.CheckboxInput($program_config['GRADES_DO_STATS_STUDENTS_PARENTS'][1]['VALUE'], 'values[PROGRAM_CONFIG][GRADES_DO_STATS_STUDENTS_PARENTS]', _('Enable Anonymous Grade Statistics for Parents and Students'), '', false, button('check'), button('x')).'</TD></TR>';
+			echo '<TR><TD>'.CheckboxInput(
+				ProgramConfig( 'grades', 'GRADES_DO_STATS_STUDENTS_PARENTS' ),
+				'values[PROGRAM_CONFIG][GRADES_DO_STATS_STUDENTS_PARENTS]',
+				_( 'Enable Anonymous Grade Statistics for Parents and Students' ),
+				'',
+				false,
+				button( 'check' ),
+				button( 'x' )
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.CheckboxInput($program_config['GRADES_DO_STATS_ADMIN_TEACHERS'][1]['VALUE'], 'values[PROGRAM_CONFIG][GRADES_DO_STATS_ADMIN_TEACHERS]', _('Enable Anonymous Grade Statistics for Administrators and Teachers'), '', false, button('check'), button('x')).'</TD></TR>';
+			echo '<TR><TD>'.CheckboxInput(
+				ProgramConfig( 'grades', 'GRADES_DO_STATS_ADMIN_TEACHERS' ),
+				'values[PROGRAM_CONFIG][GRADES_DO_STATS_ADMIN_TEACHERS]',
+				_( 'Enable Anonymous Grade Statistics for Administrators and Teachers' ),
+				'',
+				false,
+				button( 'check' ),
+				button( 'x' )
+			) . '</TD></TR>';
 
 			echo '</TABLE></FIELDSET>';
 		}
@@ -189,9 +255,21 @@ else
 
 			echo '<TR><TD>'.TextInput(Config('ATTENDANCE_FULL_DAY_MINUTES'),'values[CONFIG][ATTENDANCE_FULL_DAY_MINUTES]',_('Minutes in a Full School Day'),'maxlength=3 size=3 min=0').'</TD></TR>';
 
-			echo '<TR><TD>'.TextInput($program_config['ATTENDANCE_EDIT_DAYS_BEFORE'][1]['VALUE'],'values[PROGRAM_CONFIG][ATTENDANCE_EDIT_DAYS_BEFORE]','<SPAN style="cursor:help" class="legend-gray" title="'._('Leave the field blank to always allow').'">'._('Number of days before the school date teachers can edit attendance').'*</SPAN>','maxlength=2 size=2 min=0').'</TD></TR>';
+			echo '<TR><TD>'.TextInput(
+				ProgramConfig( 'attendance', 'ATTENDANCE_EDIT_DAYS_BEFORE' ),
+				'values[PROGRAM_CONFIG][ATTENDANCE_EDIT_DAYS_BEFORE]',
+				'<SPAN style="cursor:help" class="legend-gray" title="' . _( 'Leave the field blank to always allow' ) . '">' .
+					_( 'Number of days before the school date teachers can edit attendance' ) . '*</SPAN>',
+				'maxlength=2 size=2 min=0'
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.TextInput($program_config['ATTENDANCE_EDIT_DAYS_AFTER'][1]['VALUE'],'values[PROGRAM_CONFIG][ATTENDANCE_EDIT_DAYS_AFTER]','<SPAN style="cursor:help" class="legend-gray" title="'._('Leave the field blank to always allow').'">'._('Number of days after the school date teachers can edit attendance').'*</SPAN>','maxlength=2 size=2 min=0').'</TD></TR>';
+			echo '<TR><TD>'.TextInput(
+				ProgramConfig( 'attendance', 'ATTENDANCE_EDIT_DAYS_AFTER' ),
+				'values[PROGRAM_CONFIG][ATTENDANCE_EDIT_DAYS_AFTER]',
+				'<SPAN style="cursor:help" class="legend-gray" title="' . _( 'Leave the field blank to always allow' ) . '">' .
+					_( 'Number of days after the school date teachers can edit attendance' ) . '*</SPAN>',
+				'maxlength=2 size=2 min=0'
+			) . '</TD></TR>';
 
 			echo '</TABLE></FIELDSET>';
 		}
@@ -200,11 +278,26 @@ else
 		{
 			echo '<BR /><FIELDSET><legend>'._('Food Service').'</legend><TABLE>';
 
-			echo '<TR><TD>'.TextInput($program_config['FOOD_SERVICE_BALANCE_WARNING'][1]['VALUE'],'values[PROGRAM_CONFIG][FOOD_SERVICE_BALANCE_WARNING]',_('Food Service Balance minimum amount for warning'),'maxlength=10 size=5 required').'</TD></TR>';
+			echo '<TR><TD>'.TextInput(
+				ProgramConfig( 'food_service', 'FOOD_SERVICE_BALANCE_WARNING' ),
+				'values[PROGRAM_CONFIG][FOOD_SERVICE_BALANCE_WARNING]',
+				_( 'Food Service Balance minimum amount for warning' ),
+				'maxlength=10 size=5 required'
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.TextInput($program_config['FOOD_SERVICE_BALANCE_MINIMUM'][1]['VALUE'],'values[PROGRAM_CONFIG][FOOD_SERVICE_BALANCE_MINIMUM]',_('Food Service Balance minimum amount'),'maxlength=10 size=5 required').'</TD></TR>';
+			echo '<TR><TD>'.TextInput(
+				ProgramConfig( 'food_service', 'FOOD_SERVICE_BALANCE_MINIMUM' ),
+				'values[PROGRAM_CONFIG][FOOD_SERVICE_BALANCE_MINIMUM]',
+				_( 'Food Service Balance minimum amount' ),
+				'maxlength=10 size=5 required'
+			) . '</TD></TR>';
 
-			echo '<TR><TD>'.TextInput($program_config['FOOD_SERVICE_BALANCE_TARGET'][1]['VALUE'],'values[PROGRAM_CONFIG][FOOD_SERVICE_BALANCE_TARGET]',_('Food Service Balance target amount'),'maxlength=10 size=5 required').'</TD></TR>';
+			echo '<TR><TD>'.TextInput(
+				ProgramConfig( 'food_service', 'FOOD_SERVICE_BALANCE_TARGET' ),
+				'values[PROGRAM_CONFIG][FOOD_SERVICE_BALANCE_TARGET]',
+				_( 'Food Service Balance target amount' ),
+				'maxlength=10 size=5 required'
+			) . '</TD></TR>';
 
 			echo '</TABLE></FIELDSET>';
 		}